Start abstracting the guis
