+++ src/main/java/oop/course/client/GUI.java
@@ -1,0 +1,5 @@
+package oop.course.client;
+
+public interface GUI {
+    void startLooping();
+}
