+++ src/main/java/oop/course/client/GUIFactory.java
@@ -1,0 +1,9 @@
+package oop.course.client;
+
+import java.io.IOException;
+
+public class GUIFactory {
+    public GUI bestGUIImplementation() throws IOException {
+        return new TerminalGUI();
+    }
+}
